requirements.txt
sqlalchemy
psycopg2
python-dotenv
pandas
requests
matplotlib
seaborn
jupyter
